Feature/17 configurator service

+++ services/library/src/main/java/com/google/cloud/pso/bq_snapshot_manager/services/bq/BigQueryService.java
@@ -41,5 +41,5 @@ public interface BigQueryService {
 
     boolean tableExists(TableSpec tableSpec);
 
-    Job createSnapshot(TableId sourceTable, TableId destinationId, Integer snapshotExpirationMs);
+    Job createSnapshot(TableId sourceTable, TableId destinationId, Long snapshotExpirationMs);
 }
